Flip: fix in-game menu toggle hotkey, make watchdog ignore analog sticks

+++ spruce/scripts/gameswitcher.sh
@@ -110,8 +110,9 @@ done <$LIST_FILE
 
 # send signal USR2 to joystickinput to switch to KEYBOARD MODE
 # this allows joystick to be used as DPAD in game switcher
-killall -q -USR2 joystickinput
-
+if [ ! "$PLATFORM"="Flip" ]; then
+    killall -q -USR2 joystickinput
+fi
 # launch the switcher program
 # Usage: switcher image_list title_list [-s speed] [-b on|off] [-m on|off] [-t on|off] [-ts speed] [-n on|off] [-d command]
 # -s: scrolling speed in frames (default is 20), larger value means slower.
@@ -163,8 +164,9 @@ while : ; do
 done
 
 # send signal USR1 to joystickinput to switch to ANALOG MODE
-killall -q -USR1 joystickinput
-
+if [ ! "$PLATFORM"="Flip" ]
+    killall -q -USR1 joystickinput
+fi
 # launch game with return index
 if [ $RETURN_INDEX -gt 0 ]; then
     # get command that launches the game
